Brand Liaison..
Brand Liaison was founded in 2014 and is registered with the ROC, Ministry of Corporate Affairs, Government of India. The company provides one stop solution for all kind of compliance management services, for the Indian market. We provide compliance services to Indian and foreign manufacturers as BIS Registration, ISI Certification, BEE Registration, EPR Authorization (for E-waste and P-waste Products), WPC Approval, TEC Certification etc., which are necessary for the notified products before importing or selling in India.

The company is managed by experts who have experience in the field of compliance management and also have a technical background. As a combination of skills and leadership, Brand Liaison has established itself as one of the most trusted names in compliance management services.

Brand Liaison also provides the Brand Representation Service to foreign manufacturers, which is a necessary requirement for them in order to obtain approval and certification for the Indian market.

Brand Liaison's primary focus is helping clients to meet their legal and regulatory obligations efficiently and effectively. Brand Liaison is well-known for its cost-effective and fast processing services. We have more than 150 prominent brands among our satisfied customers, and the client list is expanding gradually.